pyproject.toml
[tool.poetry]
name = "todo api with poetry"
version = "0.1.0"
description = ""
authors = ["Kazim Hussain <87144973+kazimhussain01@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
sqlmodel = "^0.0.16"
uvicorn = "^0.28.0"
python-dotenv = "^1.0.1"
psycopg = "^3.1.18"
psycopg2 = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
